+++ src/mindagent/__init__.py
@@ -1,3 +1,0 @@
-"""mindagent public package."""
-
-__version__ = "0.2.0"